+++ propagator/src/RefSampler.h
@@ -20,13 +20,13 @@ class RefSampler
 		RefSampler(std::shared_ptr<hypercube> slow_hyper, std::shared_ptr<paramObj> par);
 		RefSampler(const std::shared_ptr<complex4DReg>& slow, std::shared_ptr<paramObj> par);
 
-		inline std::complex<float>* get_ref_slow(int iz, int iref) {return slow_ref.data() + (iref + iz*_nref_)*_nw_;}
-		inline int* get_ref_labels(int iz) { return ref_labels.data() + iz*_nw_*(_ny_+pady)*(_nx_+padx);}
+		inline std::complex<float>* get_ref_slow(size_t iz, size_t iref) {return slow_ref.data() + (iref + iz*_nref_)*_nw_;}
+		inline int* get_ref_labels(size_t iz) { return ref_labels.data() + iz*_nw_*(_ny_+pady)*(_nx_+padx);}
 
-		void sample_at_depth(std::shared_ptr<complex4DReg> slow, int iz);
-		std::future<void> sample_at_depth_async(std::shared_ptr<complex4DReg> slow, int iz);
+		void sample_at_depth(std::shared_ptr<complex4DReg> slow, size_t iz);
+		std::future<void> sample_at_depth_async(std::shared_ptr<complex4DReg> slow, size_t iz);
 
-		int _nx_, _ny_, _nref_, _nz_, _nw_, padx, pady;
+		size_t _nx_, _ny_, _nref_, _nz_, _nw_, padx, pady;
 		
 	private:
